+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Breaking
+
+  * AZ-94 Terraform 0.12 / HCL2 format
+
 # v0.1.0 - 2019-07-12
 
 Added